Formatting Camera.cpp

+++ app/src/main/cpp/Camera.cpp
@@ -25,7 +25,7 @@ Camera::Camera(int w, int h, float x, float y, float z, float ncp, float rcp, fl
 }
 
 float Camera::aspect_ratio() { // TODO: change this to a macro
-    return (float) w/h;
+    return w/h;
 }
 
 void Camera::update_width_height(int width, int height) {
@@ -38,7 +38,7 @@ void Camera::update_projection() {
     if(projection_type == PERSPECTIVE)
         perspective = glm::perspective(glm::radians(fov), Camera::aspect_ratio(), ncp, fcp);
     else
-        perspective = glm::ortho(-w/2, w/2,-h/2, h/2, ncp, fcp);
+        perspective = glm::ortho(-w/2, w/2, -h/2, h/2, ncp, fcp);
 
     cameraTranslate = glm::translate(perspective, vec3(-x, -y, -z));
 }
